GEODE-71: fix javadocs and log message
The javadocs are now correct.
Changed the log message about the host being unknown
to be warn instead of fatal since the whole point of fixing 45348
was to ignore the unknown host exception
until you try to use the pool for an operation.
I also updated the log message to say:
  "Creating pool with unknown host in case
   the host becomes known later."
instead of just saying the host was unknown.
Reviewed by: Kirk Lund and Dan Smith