GEODE-128: fix expiration tests to run faster with PR regions
Slowdown was caused by the test hook EXPIRY_UNITS_MS not being set
to true on a BucketRegion. BucketRegion does all the expiration on
a PR. The test was setting it on the bucket's parent but not no the
bucket itself. Bucket's now defer to their parent when initializing
this final field.