GEODE-133: fix testCustomEntryTtl3 to do meaningful validation and run faster
Test now validates that you can change the custom expiry action and key.
The test used to try to change the amount of time but these tests have a hard
time validating elapsed time because intermittent changes to system load and
the system clock changing due to clock sync services.