GEODE-143: logging must not cause cached values to be deserialized
A more powerful StringUtils.forceToString now exists that will also
print out the elements of arrays. It used to just to a toString on the
array which only shows the array identity. The new code will (by default)
show the first 16 element values and after that say how many more the array
had that are not shown. If you set the sys prop StringUtils.MAX_ARRAY_ELEMENTS_TO_CONVERT
you can change the default of 16.