GEODE-139: Setting local-max-memory in RebalanceOperationDUnitTest
The "coordinator" process in dunit is launched by the gradle framework.
The rest of the processes are launched by the dunit framework. Those
processes may have different heap sizes. If local-max-memory is not set,
we read the heap size to determine how to rebalance data.

In this case we didn't move any primaries because one member had much
more heap than the remaining members.