Refactoring lucene APIs into a separate module
The gemfire-core will now no longer depend on Lucene. These APIs still
have no real implementation, and there is currently also no support for
reading and writing the XML configuration as well.

As part of these changes, I modified the ExtensionPoint code slightly to
allow retrieving an extension point by key, for use in the APIs. This
lets the LuceneService APIs add the and get the LuceneService as a cache
extension.

I'm still not sure this is the best way to integrate independent modules
like this into gemfire. This needs some more discussion.