[GEODE-222] Use stable region entry snapshots in redis adapter
Although I was not able to reproduce this bug, this is most likely the fix. The encoder was given Entry objects provided by Region#entrySet which can throw EntryDestroyedExceptions if an entry becomes stale, and since the encoder was returning null upon an exception instead of letting the caller handle it, the bug only became evident when trying to hand off a null response to netty. The fix is to not catch exceptions during any encoding and to provide stable entry objects to the coder.