Fixing contents of partially written files
I added some tests of partially written files. The tests write to an
output stream but don't close it, which mimics a partially written file
because of the way the code works - it doesn't save the changes to the
File object until OutputStream is closed.

I found that the metadata was inconsistent with the file contents for
partially written files. I fixed that by deleting any chunks past the
reported length of the file.