GEODE-139 Fixing RebalanceOperationDUnitTest.testWaitFor* methods
The issue is that a previous test method was polluting the system. There
is a test of redundancy zones that sets the redundancy zone in 4 members
and disconnects the distributed system in all of those members at the
end.

However, there is a 5th member, the controller VM, that still has a
distributed system. It never had a redundancy zone set but it still
flipped a bit to indicate that redundancy zones should be enforced. From
that point on we refused to satisfy redundancy.