Adding an AsyncEventListener to write events to Lucene
Adding an implementation of the listener and a unit test. I also added
several other interfaces for collaborators with this AsyncEventListener.

SingleIndexRepository - Interface for a single lucene index (in other
words, a single index on a single bucket for a PR). This currently has
methods for writing to the index.

RepositoryManager - interface for obtaining SingleIndexRepository
objects based on a user region and a key.

ObjectToDocumentMapper - interface for translating a gemfire object
into an indexable document.