Implementing mapping objects to Lucene documents
I've added an implementation of the code that converts and object to a
document. As part of this, I renamed a couple of interfaces
ObjectToDocumentMapper -> LuceneSerializer
SingleIndexRepository -> IndexRepository

The LuceneSerializer is an interface I hope maybe we can allow users
to implement if they want so they can customize how their object gets
converted to a Document - for example for spatial indexing.

I've implemented the LuceneSerializer and added unit tests. I also
implemented IndexRepository and added a unit test for that as well.