GEODE-77 improved handling of join requests having the same host:port as a previous member
This adds a check of the UUID bytes in an address's NetMember in
compareTo to keep a new address from matching an old one.  The
test TXDistributedDUnitTest was failing consistently when a member
crashed its distributed system and then tried to reconnect immediately.
This is due to how JGroups allocates its ports starting with the first
number in the port range and sequencing through until it finds one that's
available.  GemFire had used a random selection algorithm for this.