GEODE-77 reducing aggressiveness of the health monitor
The health monitor was timing out a peer after not seeing a message from
it in 100ms.  HA testing has shown that this is far too aggressive, with
most test runs dieing a quick and tragic death with all members trying
to kick each other out of the distributed system.

This commit reverts to our old way of waiting member-timeout ms before
giving up on a peer, and adds a final check of the member activity map
before allowing a suspect message to be sent to the coordinator.