Changing the way LuceneFunctionReadPathDUnitTest uses the collector
The test should call getResult on the returned ResultCollector, not the
one provided to the FunctionService. I saw this test hang on my machine
and I'm not sure if this is the cause or not, but this is the way the
collector is supposed to be used, according to the javadocs.

Also, adding a timeout to the call so that if the test hangs again, it
will fail with a timeout.