GEODE-362: fix intermittent failures in TXExpiryJUnitTest
The TXExpiryJUnitTest now uses a test hook so it can be notified
when an ExpiryTask has expired. Note that when a tx is in progress
ExpiryTasks expire but do not perform their expiration action.
When the tx completes the expiration is rescheduled.
TxExpiryJUnitTest now also configures millisecond expiration units
so it can run faster. And the code has been cleaned up to remove
all eclipse warnings.