Hiding the lucene async event queue from the list of queues.
I added a new flag to the queue - isMetaQueue, which controls whether a
queue will show up in getAllQueues. This also has the effect of
preventing the queue from showing up in a generated xml file.

This is necessary because our queue needs to be constructed with an
async event listener that has a reference to the index. If the queue is
generated and added to the xml, it will end up trying to create a
listener before the index is created.