GEODE-77 improved suspect processing, multicast fix for Windows
GMSHealthMonitor was allowing the service to watch itself if it ran out of other members to watch.  This change set modifies that behavior to set the "next neighbor" to null, causing suspect processing to halt until a member transitions out of "suspect" state or a new membership view is installed.

I also found that GMSHealthMonitor wasn't notifying the Manager when a member goes into final-check state, so that membership listeners weren't being notified of suspect events.  This was causing a unit test failure.

JGroupsMessenger now fools the JGroups stack into thinking an IPv4 stack is being used even if IPv6 is available, so that JGroups will use an IPv4 address.  JGroups was selecting an IPv6 address, making communications with IPv4 members impossible and causing a schizm between the GemFire member address and the JGroups address.

The JGroups log level is now set to WARN in the default log4j2.xml settings file.