GEODE-77: minor bug fixes
Member weights were not being installed into member identifiers, nor serialized with them.  This wasn't an issue with GemFire because the weights were only important for JGroups IpAddresses but we now use Geode member IDs in membership views so the weights need to be included.

During startup if a member was told to become the membership coordinator it wasn't correctly setting its membership ID's view ID.  This caused some unnecessary confusion in other members.

During view preparation we don't wait for acks from members who are going to be removed in the next view but we weren't aborting the view when this happened.  This caused us to not detect a network partition.  The reply processor now adds these member IDs to the set of members that didn't respond to view preparation.

Multicast testing showed that JGroups was not dispatching a received message, probably due to problems with the message digest installed during startup.  We now send all messages out-of-band so NAKACK2 will not queue them for ordered delivery.  We also turn off bundling so that all messages are immediatly sent (no jitter for cache operations).

Other changes are in tests, which had timing issues or misspelled identifiers (this affected some constants in CliStrings, too).