GEODE-392: Removing catch Throwable from CacheTestCase.closeCache
I'm fairly certain that GEODE-392 was caused by cache.close throwing an
exception in this method. Later on, we delete the disk store files, and
then end up calling cache.close again in DistributedTestCase. The
failure we are seeing is because the disk store files are deleted but
the cache is still open.

Unfortunately, this catch throwable was swallowing all errors, so we
were not seeing failures that happened when closing the cache.