GEODE-77: proper deserialization of old member IDs
Member IDs from old GemFire clients (which we'd like to be compatible
with Geode) were not being properly deserialized, so that they did not
know what version their member was using.  This caused deserialization
errors when the IDs were transmitted from one Geode server to another.
The fix is to look for the member's version in the DataInput parameter
in fromData()