GEODE-468: Using 1.8 generated files for AnalyzeSerializablesJUnitTest
This test was passing every time with 1.8 because of a short circuit in
the test. I've regenerated the files with 1.8 and updated the test to
use junit Assume to ignore the test with the wrong JDK - at least it
will get reported as ignored.

The test was failing due to HDFS code removal. The newly generated
files also have the HDFS fields removed.