GEODE-392: Some enhancements to GemFireDeadlockDetectorDUnitTest
I originally thought this test might be the cause of this bug, so I made
a few changes to make sure it no additional threads are still running
when the test exits. I think that is still a good change to have, even
though those threads were not the cause of the issue.