GEODE-192: Removing TransactionFunctionService
This internal class was added for a use case that is no longer
supported. It has a bug, in that the client may never have connected to
the server that it supposed to be able to look up and invoke the
function on. Deleting the class rather than trying to fix the bug
because the class is no longer needed.