GEODE-522: Exclude only gradle output dirs from eclipse/idea project
Using minusConfigurations was causing gradle to exclude the entire
gemfire-core project from the gemfire-web classpath. All we wanted to
filter out was the generated output directories from gradle, which are
getting added because we have an ugly provided dependency on those
output directory.

Now we explicitly filter out only the directories we don't want.