GEODE-77: initiation of member verification on loss of shared/ordered connection
The initiation of member verification wasn't working correctly.  It wasn't
being triggered most of the time and when it was the check passed because
it was done so soon after the other member crashed that the timestamp
checks in GMSJoinLeave didn't think enough time had passed to declare the
member dead.  This changes the check to request a heartbeat first.