another attempt to fix this test's configuration
For some reason one of the build machines is causing a problem with this
test.  It should be using the same ip address for both bind-address and
locator-name but something is messing with it.  I've changed it to use
host addresses instead.  The test passes on the build machine with or
without this change when run as a single test.