fixing GEODE-588 CI Failure: LocatorDUnitTest.testMultipleMcastLocators
forcing JGroups to use an IPv4 address for multicast unless the
PreferIPv6Addresses property is set to true.  Use of multicast is
also restricted to cache operations on regions with multicast enabled.
This avoids problems when multicast isn't properly set up and a user
enables use of multicast in the DistributedSystem config.