Modifying the fix for handling old client IDs
The previous fix for handling old client identifiers left the IDs always
thinking that the client was running version GFE_82.  The new fix makes
the client defer to its encapsulated InternalDistributedMember identifier
to determine the client's version and handles deserialization of the
encapsulated pre-GFE_90 identifier in InternalDistributedMember.fromData().

This fix has the added benefit of reducing the size of client identifiers.