GEODE-709: CI Failure in AnalyzeSerializablesJUnitTest
JDK 1.7 added three new types of entries to the "constant pool" of
a compiled class: InvokeDynamic, MethodHandle and MethodType.  The test
was failing when it ran into an InvokeDynamic entry (type code 18) in some
compiled class.

Evidently we had no classes that required an InvokeDynamic entry until
recently.

I also changed the test so it will log the name of the class that causes
a failure in constant-pool parsing in the future.