Fix for suspect string causing a number of CI failures
During shutdown it's possible for the DirectChannel communication service
to throw a ConnectExceptions containing a ShunnedMemberException for one
or more members.  These indicate that the message couldn't be sent, but this
is expected behavior and the ShunnedMemberExceptions can be ignored.
This fixes GEODE-355, GEODE-682, GEODE-683, GEODE-684, GEODE-685, GEODE-687,
GEODE-689, GEODE-691, GEODE-702 and GEODE-703.