merge of hot-fix from closed-source 8.2x maintenance branch
The tombstone GC service ran into an NPE removing a tombstone and then
the thread went hot because exception-handling was not advancing to the
next GC candidate.  This revision attempts to avoid the NPE (which was
not reported with line numbers) and alters the exception-handling to
advance to the next tombstone if there is an unexpected problem removing
the current tombstone.