fixing CI failure
This test periodically failed to find 2 live servers due to an
assumption in the test that Pool.borrowConnection() would never find
the initial connection created by the pool in prefillConnections().
That caused it to only be connected to one of the three servers instead
of the expected two.  The fix is to explicitely ask for a connection
to the second server.