GEODE_368: CI failure in HARQueueNewImplDUnitTest.testConcurrentGIIAndDispatch
removed an old assertion that is no longer valid and rewrote the code to
allow for the possibility of a null entry in the queue.

This commit also includes renaming of one of the old BugXXX unit test classes
that I'm currently working on to have a name that matches what it is testing.